feat: update content and add avg
