+++ pages/stats.tsx
@@ -1,11 +1,9 @@
 import { GetStaticProps } from 'next';
 import { clientMatomoStats, IMatomoStats } from '#clients/matomo';
-import BasicChart from '#components/chart/basic';
 import Meta from '#components/meta';
-import { ScoreNpsStats } from '#components/stats/score-nps';
+import { NpsStats } from '#components/stats/nps';
 import { TraficStats } from '#components/stats/trafic';
 import { UsageStats } from '#components/stats/usage';
-import constants from '#models/constants';
 import { NextPageWithLayout } from './_app';
 
 const StatsPage: NextPageWithLayout<IMatomoStats> = ({
@@ -22,12 +20,6 @@ const StatsPage: NextPageWithLayout<IMatomoStats> = ({
       title="Statistiques d’utilisation de l’Annuaire des Entreprises"
       noIndex={true}
     />
-    <script
-      src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.7.0/chart.min.js"
-      integrity="sha512-TW5s0IT/IppJtu76UbysrBH9Hy/5X41OTAbQuffZFU6lQ1rdcLHzpU5BzVvr/YFykoiMYZVWlr/PX1mDcfM9Qg=="
-      crossOrigin="anonymous"
-      referrerPolicy="no-referrer"
-    ></script>
     <h1>Statistiques d’utilisation</h1>
     <p>
       Découvrez nos statistiques d’utilisation mises à jour quotidiennement.
@@ -44,32 +36,9 @@ const StatsPage: NextPageWithLayout<IMatomoStats> = ({
       mostCopied={mostCopied}
     />
     <h2>Satisfaction des utilisateurs</h2>
-    Analyse des réponses au{' '}
-    <a href="/formulaire/nps">questionnaire de statisfaction</a>.
-    <h3>Score de satisfaction</h3>
-    Ce qui nous intéresse avec ce score n’est pas sa valeur, mais ses
-    variations. Si le score augmente, c’est que le service évolue dans le bon
-    sens et inversement.
-    <ScoreNpsStats
+    <NpsStats
       monthlyAgentNps={monthlyAgentNps}
       monthlyUserNps={monthlyUserNps}
-    />
-    <h3>Répartition des réponses par catégorie d’utilisateurs</h3>
-    <BasicChart
-      yLabel="Répartition"
-      stacked={true}
-      horizontal={true}
-      yRange={[0, 100]}
-      height={50}
-      type="bar"
-      labels={['']}
-      datasets={Object.keys(userResponses).map((userTypeKey, index) => {
-        return {
-          label: userTypeKey,
-          data: [userResponses[userTypeKey].value],
-          backgroundColor: constants.chartColors[index + 1],
-        };
-      })}
     />
   </>
 );
